Partially implemented track change notifications (title/artist).
The client is now notified when the title or artist of a track in the collection changes. This usually happens during the first full indexation. When this happens, the music collection list in the GUI remote is updated.

Still TODO: send notifications when only the availability of a track changes.