The server now requires the client to be authenticated before allowing commands such as "play".
Several types of client requests are now ignored or answered with an error message if the client has not yet successfully completed a user login procedure. This means that the commandline remote can no longer be used to control the server for now, because it does not support authentication yet. That will need to be fixed, although I must admit it's not very high on my priority list.