Fix the hash calculation, but also keep the legacy calculation around.
The legacy hash is still calculated, so we can link it to existing data in the track
listening history. The new hash will be used if the legacy hash is not yet present in the
database. In the near future, we will need a mechanism to add both to the database and to
link them together.
The FileData class has been removed, and a new class FileAnalyzer has been created.