More extensive unit testing for the hash.
In the unit test we are now doing several modifications (in a memory buffer) to the test file. We check if the modifications are real, we calculate the hash for each of them, and we check if the resulting hashes are all the same, the one we expect it to be.

But more types of modifications are still needed. We need to try to change as many individual tag fields as possible, for all kinds of tag types (ID3v1, ID3v2, ...), and then see if all those still result in the hash value we expect.