More extensive unit testing of the hash.
We now test the hash even more by generating many different tag modifications to the test file and checking if they all result in the same hash value.
Unfortunately this reveals a problem with the hashing of MP3 files; apparently APE tags are not removed correctly before calculating the hash value, resulting in a hash difference.
Hashing of FLAC files seems to be working correctly for now.