Test the hash algorithms more extensively.
We now try all possible combinations of two consecutive modifications to our in-memory copy of the file, and check if this still results in the hash value we expect.

The hash algorithms had to be modified slightly; they will now anticipate the presence of up to *two* ID3v1 tags and strip them both before calculating the hash value. This was necessary because all TagLib versions < 1.11.1 fail to detect tags at the end of a file when using an in-memory file buffer instead of a 'real' file as input. The consecutive modifications we do to the in-memory test file cause such older TagLib versions to generate two ID3v1 tags, one for each modification, because at the start of each modification it has to parse the file again.

The modification to the hash algorithms is not expected to have a big impact; it will only affect real music files on disk that already have two or more ID3v1 tags, which should be very rare. Such files would be assigned a different hash value from the one they got before this commit. I doubt if anyone besides me actually uses PMP at this time, and if there are, the odds of any of them being affected by this change are very small, so I'm not going to provide a workaround.