Change server logging and add some TODO's.
The server now uses our custom logging mechanism, though it only logs to stdout for now. The logging to console has been secured with a mutex, so different threads writing to the logs at the same time will no longer result in unreadable text. I intend to switch the server to log-to-file soon, like the client is already doing right now.

Added a new section "PLANNED NEXT" to the "TODO.txt" file. This is supposed to be a reminder of what to work on, not a promise of any kind.