Add recent player history view to the GUI remote.
A list of recently played tracks is displayed in the GUI client, above the playback controls. The idea is that a track first appears in the queue below, gradually moves up to the top of the queue, then goes into the player (the playback controls), and comes out again and appears right above the playback controls, on the bottom of the recent playback history.

The history view is a very rough first version, it needs further polishing. It is currently not resizable yet, and it does not automatically scroll to the bottom of the list yet when a track is added. And more (useful) information about the tracks could be displayed in the list.