Automatic scroll to most recent track in the recent history list.
The list is scrolled to the bottom every time a new entry is added. The track most recently played is always added at the bottom of course.
The user can scroll up again until another item is added.

This change also has the nice side-effect that the most recently played track is immediately visible when the client is started.