Set main window title explicitly.
We now set the window title explicitly for our main window. Until now Qt automatically used the application name as window title. That was fine to me, but after the program had been running for a while and hibernated/suspended a few times, the window title somehow changed into the name of the executable minus the extension. Which was not so pretty. So let's hope that setting the window title explicitly with a setWindowTitle() call will fix this.