Add a safer switch to binary mode
The text command 'binary' that the client sends to the server in order to ask a switch to binary communication, is not supposed to be used by a person who is typing commands to control the server. So we add support for an extra argument to the command, a fixed string that is (nearly) impossible to remember. The old form, without the argument, is still accepted by the server for compatibility reasons, but it will not support this forever. The client now (mis)uses the greeting text of the server to detect if it can use the new style 'binary' command or not. It will use the new style if possible, and otherwise fall back to the old style command without argument. Ten years from now, or hopefully much sooner, support for the old style will be removed completely.