Change and extend sorting of the music collection
Sorting of the music collection is now done in our own implementation (SortedCollectionTableModel) instead of in a class derived from QSortFilterProxyModel, because the latter could not do correct multi-column sorting. We still use it for filtering though.
Added sorting implementations for sorting by track length and album name.
When sorting on track title or track artist, the tracks with both title and artist unknown are from now on always put on the bottom on the list; previously they were on the bottom only when the list was sorted ascending, but they were on top of the list when the sorting was done descending. This is an experiment; we'll see how this affects user experience (there's a good chance I am still the sole user).