Create a table in the database for file sizes
Make sure a database table is created that will be used to store the possible file sizes for each hash. We will be able to use this for quickly locating files that match a specific hash when we don't know the correct file name to look for; for example when looking for a file that has been renamed since the last full indexation. It can also be useful to avoid having to run a full indexation at server startup; all hashes could be speculatively matched to files on the file system based on the information in our database tables of file sizes and file names.

The logic to add any records to this table has not been implemented yet.