Fix Util::getRandomSeed and expand its unit test
The unit test of Util::getRandomSeed failed for me today, which it had never done before and also without anything having changed in its implementation. The failure can be explained by getRandomSeed being called twice by the unit test with almost no time in between. It's likely that the system clock increased by 1 millisecond while the counter also increased by 1. This means that the clock and the counter probably cancelled each other out.

I fixed this by putting a very short QThread::msleep() call inside Util::getRandomSeed(). This will decrease the probability of such a coincidence happening again. A call to getRandomSeed() will now take a bit longer, but the function will not be called very often, so it does not really matter.