Check if the preloader cache file still exists
In the preloader, check if the preloaded file still exists before trying to use it. It may have disappeared because preloaded files are stored in a TEMP directory. In case it has disappeared, update our cache status accordingly and then pretend to the Player that the file wasn't preloaded at all.