Fix hash-fetching of queue entries (breakpoint)
Fetching the hashes of the queue-entries did not work when there was a breakpoint in the queue. This resulted in the "Last heard" and "Score" columns remaining empty in the client.

The cause of this problem was the server generating an invalid message. The problem in server has now been fixed and no modification to the client was (really) necessary. Older clients will also benefit from this fix.