Add "Start high-scored tracks wave" action
This action will put dynamic mode into a special mode of operation where it will only generate tracks with a high score. The minimum score a track must have changes after each track that is generated; first it goes up until 85% and then it goes back down. When the minimum score has dropped enough, the action is considered to be over and can be restarted by the user if desired.