Add the ability to duplicate a queue entry
Add "duplicate" option to the queue context menu in the GUI remote. This will create a copy of the track (or breakpoint) and insert it right after the original. It provides an easy way to play a song multiple times, not having to drag it from the music collection to the queue over and over.

Introduce the ServerConnection::ResultHandler class for dealing with result/error messages in a better, more structured way.