GUI-Remote: hide Win32 console for Debug builds
Until now, a console window was shown if the build was in Debug configuration or if the ENABLE_WIN32_CONSOLE configuration variable was set.
From now on, only look at ENABLE_WIN32_CONSOLE. This means that a console window will no longer be displayed by default, as the ENABLE_WIN32_CONSOLE variable defaults to off.