Use windeployqt for copying the Qt dependencies
The Qt directory is no longer specified explicitly, but it is assumed that the Qt "bin" directory is present in the Windows PATH variable. So we invoke windeployqt without specifying its path.

The MinGW toolkit directory is still explicitly listed in the script, because we need it in order to find the TagLib DLL (libtag.dll).