Cmd-Remote: add support for server shutdown again
For a while now, the command-line remote was not able to shut down the server anymore, because the server rejects all commands that have side-effects as long as no user authentication has been done. And the command-line remote does not support user authentication yet.

This commit provides a way to shut down the server using the command-line remote once again, by requiring the server password as an argument. This proves that the user initiating the shutdown has control over the server, because the server password can only be observed from the stdout of the server. Exposing the server password as a command-line argument is acceptable, because the password instantly becomes useless after the server is shut down. The server generates a new server password each time it is started.