Wait a bit before trying to scrobble
Wait a bit before trying to scrobble the track that has just finished playing. This gives the database enough time to store the new history entry. Before this change, it sometimes happened that the history entry wasn't yet returned from the database query that determines the tracks to be scrobbled.