Send track availability change notifications
When a track becomes available/unavailable, send a message to the client. Make sure the client processes the notification correctly.

Migrate some uses of QList to QVector.

NOTE: track availability changes are currently only detected by the server in some scenarios, not always. Running a full indexation will only detect new tracks, not tracks that have disappeared. But track disappearances/reappearances will be detected for tracks in the queue when they are not too far down. A track that has disappeared will be shown in a gray color in the music collection view.