Allow the client to enable or disable scrobbling
Extend ScrobblingHost:
 - prepare for other scrobbling providers in the future
 - add support for enabling/disabling a scrobbling provider for some user
 - add support for relaying scrobbler status to the client

Add new top-level menu "User" to the GUI client:
 User > Scrobbling > Last.FM

The "Last.FM" menu entry is checkable and is used to enable or disable Last.FM scrobbling for the current user. Disabling scrobbling will not prevent any tracks that are played from being scrobbled in the future when scrobbling is reenabled.