Add option to prevent display turning off
Add an option "Keep display active during playback" to the View menu of the GUI remote.
This option, when enabled, prevents the display from being turned off automatically when the system is idle for some time. It is a local option, just for the client, and does not affect the server. The option only has effect during playback, when the player is not "stopped" or "paused".

Implementing this required platform-specific code because Qt does not provide this kind of functionality yet. For now only Windows is supported, and the option is disabled on all other platforms.