Add a simple foundation for protocol extensions
Add a mechanism that will make it possible to extend the network protocol in parallel, ensuring that extensions to the network protocol in development branch A will not conflict with other extensions in development branch B when a client of type A is talking to a server of type B or vice versa.

I intend to start using this mechanism in the scrobbling branch.