Fetch track scores all at once
For the track highlighting feature we need the score for each track. Until now, the score was only fetched from the server when it was needed, like when scrolling through the collection view while doing track highlighting based on score. This did not provide a good user experience, as the highlighting often did not appear until scrolling past a track. So we now make sure to start fetching the scores immediately, before the user even enables track highlighting.