Mark the current track in the collection view
The collection view will now show a marker to the left of the title of the track that is currently playing. The marker will also indicate if the track is playing or paused.

For now this is done using simple Unicode symbols, but the idea is to switch to icons in the future.