Fix automatic switch to personal mode
The server always starts up in public mode. When a client connects to that server it will then automatically switch the server to personal mode and also enable dynamic mode, if the server is not playing anything.

This automatic switch was broken by recent refactorings. This commit fixes it and extends ClientServerInterface a little further.