New class for obfuscating third-party credentials
Sometimes we need to transmit a password as-is from the client to the server,
like for activating scrobbling to Last.FM. Because our client-server
communication is not encrypted at this time, we will need some kind of
obfuscation.

I expect that transmitting credentials this way should be rare. For Last.fm we
would only need to do it the first time the user activates scrobbling. We will
use the credentials to obtain an access token that will be stored in the
server's database. After that we do not need the user's credentials anymore
until the access token is no longer accepted by Last.fm.

This commit only adds the new class, but it isn't used anywhere yet.