Send track length in milliseconds to the client
When the client requests information about what tracks are in the queue, we
send information like title, artist and track length. Until now, track length
was transmitted as a number of seconds. This commit changes that. Track
length is now transmitted as a number of milliseconds. The client now takes
advantage of the extra precision and displays track length in short form like
"03:44.7" or in long form like "00:03:44.726".

The extra precision this commit introduces is only related to the queue, the
track that is playing, and the history view. The music collection view and
the track info dialog already had access to track length in millisecond
precision but did not take advantage of this yet.

Backward compatibility:

Clients that have this improvement can still connect to an old server that
does not have this improvement and vice versa. But when a new client connects
to an old server, the user may notice that track lengths differ between the
the queue and the music collection view, for example.