CMD Remote: introduce user authentication (WIP)
This is the first part of a fundamental rewrite of the command-line remote.
The tool is partially broken now, but it was already broken before, now it's just broken in a different way.

This commit introduces user authentication. The remote will now ask for a username and password. Authentication is required for all commands that change the state of the server, like play, shutdown etc.

Commands that are functional:
 - play
 - pause
 - skip
 - shutdown (now without arguments, see below)
 - volume (changing the volume)

Commands that are *not* functional:
 - nowplaying
 - queue
 - volume (getting the current volume)
 - qmove

The "shutdown" command that instructs the server to exit now no longer accepts any arguments. Until now it required the server password as an argument, but it no longer does.

Most commands that instruct the server to perform some action now wait for confirmation from the server that the action was successful. The "shutdown" command is the only exception, because the server does not have any kind of confirmation yet for a shutdown request.

The 'usage' help text of the command-line remote has not been updated yet, so it is largely out-of-date.

This commit switches the command-line remote from the archaic text-based communication protocol to the regular binary communication protocol that is also used by the GUI remote.