Save & restore dynamic mode preferences per user
The server will now store the dynamic mode preferences for each user in its
database, and will load those preferences when switching to that user.

Those preferences currently include whether or not dynamic mode is enabled
or not, and the "avoid repeating the same track" interval.

The client will now no longer enable dynamic mode automatically when
connecting to a server, because this is now the server's job when loading
dynamic mode preferences for the user. This means that when a new client
connects to an older server, dynamic mode will no longer be enabled at
logon. And an old client that connects to a new server will overwrite the
user's preferences by always turning on dynamic mode at logon.