CMD Remote: reimplement the "queue" command
This makes the command-line remote fully functional again in the sense that
all original commands are working again.

The output of the "queue" command has changed a little bit.

For now, the "queue" command requires the user to authenticate. This is a
bit unfortunate, but there is no way around that because the server
currently requires authentication for downloading the queue. Fixing this
will require a change in the server, meaning that the new remote would no
longer be compatible with older servers unless I implement a potentially
complicated workaround.