CMD Remote: add "login" command (not useful yet)
The "login" command is used as an extra command that is specified before
the actual command you want to execute. It does not take any arguments yet.
It currently only forces authentication to occur, even if the other command
specified does not need authentication. The plan is to eventually use this
command for non-interactive authentication.

Example usage:
  PMP-Cmd-Remote localhost login : nowplaying

The colon character is used as a separator between the two commands. The
spaces next to the colon can be omitted.