CMD Remote: add non-interactive authentication
Add support for reading username and password from standard input. This
makes it possible to use the command-line remote from a script.

Both username and password can be read from standard input, or just the
password if the username is specified as a command-line argument.

When username and password are both read from standard input, it is
assumed that the first line of the input is the username and the second
line is the password.

Example usage:
  PMP-Cmd-Remote localhost login - : play <credentialsfile
  PMP-Cmd-Remote localhost login Username - : play <passwordfile