Add ability to find files that have moved
The server can now find a file that has moved, even when its name has
changed. It does a quick scan for new files (that are not indexed) and
looks at those. Files having a size that matches a previously known file
for the same hash are examined first.

Note that the server only does this kind of search in some cases: only
when it needs to (pre)load the file or when it expects it will have to
load it soon (when the track is near the top of the queue).