Add ability to display time remaining for track
Clicking on "position" will change it to "remaining" and display the time
remaining for the track. Clicking again will switch back to "position".

This change reveals broken HiDPI handling of the old Qt version that I am
still using. The text "position" and "remaining" is clipped at the end when
using display scaling. Moving the window to another screen and back fixes
the problem for me. I hope this problem will disappear when upgrading Qt to
a more recent version.