[WIP] Compatibility UI, part 1
The compatibility user interfaces mechanism will allow an older client to
access some of the newer features of a newer server.

For each new feature that the older client does not support, the server
will expose a very primitive UI over the network to control that feature.
It will be nothing more than some text to display and a few actions that
are to be displayed as push buttons.

For example: if the server gets support for timer based auto-play in the
future, it can export a UI over the network that notifies the user that a
countdown is in progress and shows a button that allows the user to cancel
the countdown.

This first commit implements most of the network protocol changes, but not
much more than that.