Add support for reloading the server settings
Server settings can now be reloaded at runtime. For now this is only useful
for changing the music paths setting without having to restart the server.

The reload can be triggered by the command-line client with the new command
"reloadserversettings". The GUI client does not support this new feature
yet.