GUI Remote: add button "Scan for servers"
Add a button that allows the user to re-run server discovery.

This can be useful when the user starts the client and then realizes that
the server machine is sleeping and is not auto-detected. Without the new
button the user had to restart the client after waking up the server
machine, or connect to the server manually using its hostname or IP.