Add support for inserting breakpoints at any index
Until now it was only possible to insert a breakpoint at the front of the
queue, and only if there was no breakpoint present yet at that location.

This change adds support for inserting a breakpoint at any location in the
queue and does not depend on the absence of an existing breakpoint at the
same location.

For now only the command-line remote has support for this new feature.