Add support for adding barriers to the queue
A barrier acts like a break, but is not removed from the queue by the
player.

For now only the command-line client has support for inserting barriers
into the queue. But the desktop client can display them if present.

Barriers are useful when you want to take a break from random track
selection and select a few tracks manually: you insert a barrier at the
front of the queue and then insert one or more tracks before the barrier.
Playback will always stop at the barrier, and each time playback stops you
can decide to add some more tracks yourself and play those, or to delete
the barrier and resume playing random tracks.