Disconnect when client references about to run out
'Client reference' numbers are not reusable, so they get bigger and bigger.
We don't want the number to overflow, so we now break the connection to the
server before this can happen, when the most significant bit of the number
becomes 1.

In practice it is highly unlikely that a client-side connection lives long
enough so that this new safety mechanism is triggered.

In the future I would like to have the client do a smooth transition from
an old connection to a new connection when it is almost time to close the
old connection.