Implement client-side connection keepalive
Introduce keepalive messages. The client will now send a keepalive message
if it did not receive anything from the server in a while. The server must
then respond with a keepalive message of its own. If the server does not
respond within a few seconds, the client will assume that the server is no
longer connected and it will terminate the connection.

This also works when the client is connected to an older server that does
not know about keepalive. In that case the client will send another kind of
message that requires a response from the server. Currently this is a
request for the status of dynamic mode, but this may change in the future.