Treat hashes of the same file as equivalent
A file can have multiple hashes when analyzed: a regular hash, but
sometimes also a legacy hash. Both hashes may have been used in the past
for the playback history of that file. From now on we take the history of
all hashes of a file into account when calculating its score and
'previously heard'.