Update code for start-up position of main window
Update the workaround for QTBUG-77385, rewrite the code to get rid of
the deprecation warnings.

Change the repositioning logic: resize the window to 80% of the
available screen geometry (previously 50%), and now take the taskbar
into account for calculating the new position.

Another improvement: when detecting that screen geometry was not really
restored from a previous run (when running the application for the first
time or when restore failed), apply the repositioning logic instead of
leaving the window as is.