Music collection view: add second filter field
Add a second filter combobox to the music collection view. The second
filter field will only be visible when the first filter is set to
something other than "none". Both filters are combined with a logical
"AND" operation, so a track must satisfy both filter criteria in order
to be shown in the list.