"nowplaying" cannot easily print album
The "nowplaying" command cannot simply print the album of the track
like it already prints title and artist, because album information is
not available directly. In order to get the album, we would have to
get track information via the hash instead of the queue ID.

This commit adds two commented lines for printing the album that do not
work (yet).