Add track filter "no longer available"
This new filter criterium lists tracks that have disappeared from disk
since the server was started. This means that the server has realized
that the track's file no longer exists and an alternative file has not
been found yet.

NOTE: the server can find tracks that have been moved or renamed on
disk, but this currently does not happen automatically. The server will
only do this when the track is in the queue, near the front.