Implement client-server scrobbling authentication
Initial basic implementation of scrobbling authentication. It allows the
user to authenticate with Last.fm, so tracks listened to can be
scrobbled. In this initial implementation only the command-line client
can be used for authenticating. The desktop client will get support for
this in the future.

Run the command-line client with "--help" to get more information about
the "scrobbling" command and how it can be used.

Note: passing e-mail address as username does not work at this time.