Add user statistics display switch
This allows switching the display between:
 * personal statistics
 * public statistics

The switch affects the statistics (score and last heard) shown in the
queue, the statistics used for the filters and highlighting in the music
collection, and the user account that is selected by default when
opening the track info dialog.

When switching between public mode and personal mode, the new switch
will adjust itself to match the new mode. When switching to public mode,
the display switch will be set to display public statistics, and vice
versa. But the display switch will not have any effect on whether public
mode or personal mode is active.

This new switch makes public mode more usable.

Note: dynamic mode is not affected by this new switch. Dynamic mode is
still based on whether public mode or personal mode is active.