Music collection view: add third filter field
Add a third filter combobox to the music collection view. The third
filter field will only be visible when the first or the second filter is
set to something other than "none". All filters are combined with a
logical "AND" operation, so a track will only be shown in the list if it
satisfies all of the filters that are not set to "none".