Remove delay for starting the scrobbling mechanism
Start the scrobbling mechanism immediately at startup of the server,
instead of waiting until the first full indexation has finished.

We can now do this because TrackInfoProviderImpl will first locate a
file for the track to be scrobbled, and as a side effect information
about the track that we need for scrobbling, like title and artist, will
be loaded from that file. This will not work for all cases, but it is
good enough because the probability that the server needs to do
scrobbles at startup is low. When this approach does not work, meaning
that if file is not found easily or if the first file found does not
contain title and artist of the track, the scrobble will be skipped.