Take hash equivalences into account
When fetching the history of an individual track, take hash equivalences
into account. This means that we will return history entries of all
hashes that are considered to be equivalent to the hash that was
requested.

This is a change in the server only, so older clients will also benefit
from this improvement.