Add a way to get track info of an individual track
Extend the network protocol in order to make it possible to get title,
artist, length, and other properties of a single track. Until now it
was only possible to obtain this information by downloading the entire
list of tracks from the music collection to the client.

Add the "trackinfo" command to the command-line client. This command
prints title, artist, album, album artist, length, and availability of
a single track.

The track info dialog of the desktop client can now display the
information of a track before the entire collection has been downloaded
to the client.