Music collection view: add search by track hash
Pasting a track hash into the search field will now search for the track
with that particular hash. This is an exact search; pasting part of a
hash will not work and will do a regular textual search instead.

Since a hash uniquely identifies a track, there will never be more than
one result when searching by hash. There can be zero results if one of
the following conditions is true:

 - one of the other filters excludes the track from the result
 - the client is still downloading the list of tracks from the server
 - the server is not familiar (anymore) with the track
 - you are using a legacy hash value (an old alias of the real hash)

The hash of a track can be obtained using the "copy hash" button of the
track info dialog or using the command-line hash tool.