Send hash lookup request to the server if unknown
When searching for a hash that is unknown to the client, send a hash
lookup request to the server in the hopes that the server does know the
track. This small change fixes the (potential) long wait time when the
client is still downloading the list of tracks from the server.