Introduce class UserHashStatsCacheFixer
The new class is meant to fix the cache when we detect that the
database has been touched by an older server without properly updating
the cache.

Also revise the way scores are calculated/invalidated. Distinguish
between invalidating the score of an individual hash and invalidating
the score of an equivalency group.