Change the play/pause/skip button icons
Add new SVG icons for the play/pause/skip buttons. The new icons have
the same color as the app icon.

Decrease the size of the buttons a little bit.

The old icons are not removed yet. The old play/pause icons are still
used by the music collection view.