Use "--:--:--.---" for unknown track length/posit.
When there is no known track length or track position available,
display "--:--:--.---" instead of an empty string. Usually this is when
the player is in the stopped state.

We use figure dashes instead of regular dashes, to ensure that the new
placeholder text has the same width as an actual track length or
position that consists of actual digits.