Handle MySQL error 2013 and retry the SQL query
It seems we now receive MySQL error 2013 instead of 2006 if the database
connection to the server was broken somehow. But error 2013 could also
indicate a query timeout, so we measure the elapsed time between start
of the query and the error and only retry the query if the elapsed time
was very short.

Make the executeQuery function a bit shorter by moving some things into
separate functions.