Fix scrobbling of tracks before initial indexation
Make sure that tracks can be scrobbled before the initial indexation is
run or completed. This is necessary at server startup when faced with a
backlog of tracks to scrobble.

We already had a mechanism for this, but it needed improvement.

Our strategy:

 1. try to find a file for the track that we need to scrobble
 2. wait until Resolver has processed the file's data (NEW STEP!)
 3. get track info (title, artist...) from Resolver
 4. do the scrobble if we have at least title and artist

In step 1 we try to locate a file for the hash. The server will look at
certain files based on some heuristics (file name and file size),
analyze those, and will return the file name if the file analysis
reveals that it is the hash we were looking for.

Step 2 is new, and is added by this commit. We need to wait until the
Resolver class has integrated the results of the file analysis from
step 1 into its data structures. Without step 2 there was a race
condition present: between finding the file and asking Resolver for
artist and title there could be a brief period where Resolver has not
processed the result of the file analysis yet and we could end up at
step 4 with title and artist still missing.