Create 'maintenance' folder and move 4 files there
Move the following files into 'maintenance':

 - RELEASE PROCEDURE.md
 - TODOs.txt
 - create-x64-windows-release-zip.cmd
 - sourcecount.sh

Adjust the two script files because they can no longer assume that they
are located in the top-level directory.

Fix a bug in sourcecount.sh: the script was still looking for the
'gui-remote' folder which no longer exists since the GUI Remote was
renamed to Desktop Remote.