Add script for a PMP debug build on Windows-x64
This new script builds PMP and copies all runtime dependencies to the
build folder, so the PMP executables can simply be run by double clicking
them in Windows Explorer.

The script creates a separate build directory for this instead if the
standard "bin" directory.

There is still something going wrong inside the script. The invocation of
windeployqt seems to terminate the script, and everything after that
point is not executed anymore.