Import dark color scheme (needs update)
PMP has some extra colors it defines itself, like the colors for
highlighting tracks in the music collection. The develop branch only has
a light color scheme for these colors. This commit adds a dark color
scheme, copied from the old "dark theme" implementation.

From now on, we will use either the light or the dark color scheme,
depending on the value returned by QStyleHints::colorScheme(). This
means that we will now be able to support both light *and* dark mode.

The colors of the dark color scheme will need to be changed because they
were defined to match the colors of the dark style sheet we were using
in the old implementation, and they do not match the colors of the Qt
fusion theme.