Set relative font size in code for artist/title
Dark mode in PMP is affected by QTBUG-112966. Labels that use style
sheets don't change color automatically when the system switches from
light mode to dark mode or the other way around while the app is
running. So we will need to avoid using style sheets.

This is a first change to replace a style sheet with some code that does
the equivalent thing. Many usages of style sheets are still remaining.

The style sheet we eliminate here was using an absolute font size, while
the new bit of code sets a relative font size. So this is another
improvement.