Queue: fix missing refresh of the last two columns
Because of off-by-one errors, the last two columns of the queue were no
longer being refreshed when necessary. This was most obvious right
after login: the columns stayed empty until forced to refresh by
dragging the divider between the queue and the music collection.

This problem appeared because of the migration to Qt 6. It seems that
Qt 5 was not that strict about model indexes being out of range, while
Qt 6 refuses to invalidate the model when indexes are out of range.