Add support for adding labels to a track
This introduces labels and the ability to apply a label to a track.

The name of a label must be in lowercase. Allowed characters are
letters, numbers, and dashes. Dashes are not allowed at the start or
the end of a name, and multiple consecutive dashes are not allowed
either.

The first support for applying a label to a track is found in the
command-line remote only. The new command is:

  label add <name> <hash>